Renaming components and containers to a better convention
